feat: initial commit
skeleton build for the project
need to think more about how the modules are going to look
i think it makes sense that each chapter is a module, and i can add many
functions that implement the ideas
then i can add test suites that implement the respective chapters

as a side note, i can't figure out parallel policies for some reason. i
am not sure if that's because i'm using clang and it's just not
supported. it's supported for gcc though??